Elasticsearch: rename public/app/plugins/source/elasticsearch/query_def.ts to follow convention (#55415)
