+++ public/app/plugins/datasource/elasticsearch/hooks/useFields.test.tsx
@@ -8,7 +8,7 @@ import { BucketAggregationType } from '../components/QueryEditor/BucketAggregati
 import { ElasticsearchProvider } from '../components/QueryEditor/ElasticsearchQueryContext';
 import { MetricAggregationType } from '../components/QueryEditor/MetricAggregationsEditor/aggregations';
 import { ElasticDatasource } from '../datasource';
-import { defaultBucketAgg, defaultMetricAgg } from '../query_def';
+import { defaultBucketAgg, defaultMetricAgg } from '../queryDef';
 import { ElasticsearchQuery } from '../types';
 
 import { useFields } from './useFields';
