+++ public/app/plugins/datasource/elasticsearch/specs/elastic_response.test.ts
@@ -2,7 +2,7 @@ import { DataFrameView, FieldCache, KeyValue, MutableDataFrame } from '@grafana/
 import flatten from 'app/core/utils/flatten';
 
 import { ElasticResponse } from '../elastic_response';
-import { highlightTags } from '../query_def';
+import { highlightTags } from '../queryDef';
 import { ElasticsearchQuery } from '../types';
 
 describe('ElasticResponse', () => {
